pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "spire_fyi"
version = "0.1.0"
description = "spire.fyi utility package"
readme = "README.md"
authors = ["spire_fyi <info@spire.fyi>"]
license = "MIT"
repository = "https://github.com/spire_fyi/spire_fyi"
homepage = "https://github.com/spire_fyi/spire_fyi"

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  #! Update me

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  #! Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
]

[tool.poetry.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
# "spire_fyi" = "spire_fyi.__main__:app"  #  #TODO: add query_data CLI

[tool.poetry.dependencies]
python = "^3.10"

typer = {extras = ["all"], version = "^0.4.0"}
rich = "^10.14.0"
pandas = "^1.5.0"
numpy = "^1.23.4"
scipy = "^1.9.2"
shroomdk = "^1.0.1"
altair = "^4.2.0"
jinja2 = "^3.1.2"
pillow = "^9.2.0"
requests = "^2.28.1"
scikit-learn = "^1.1.2"
streamlit = "^1.13.0"
networkx = "^2.8.7"
matplotlib = "^3.6.1"
pyvis = "^0.3.0"
helius = "^0.0.3"
solana = "^0.29.1"
solders = "^0.14.4"
anchorpy = {extras = ["cli"], version = "^0.16.0"}

[tool.poetry.group.dev.dependencies]
ipython = "^8.5.0"
jupyterlab = "^3.4.8"
jupyter = "^1.0.0"
jupyter-core = "^4.11.2"
bandit = "^1.7.1"
black = {extras = ["jupyter"], version = "^22.10.0"}
darglint = "^1.8.1"
isort = {extras = ["colors"], version = "^5.10.1"}
mypy = "^0.910"
mypy-extensions = "^0.4.3"
pre-commit = "^2.15.0"
pydocstyle = "^6.1.1"
pylint = "^2.11.1"
pytest = "^7.2.0"
pyupgrade = "^2.29.1"
safety = "^1.10.3"
coverage = "^6.1.2"
coverage-badge = "^1.1.0"
pytest-html = "^3.1.1"
pytest-cov = "^3.0.0"

[tool.black]
# https://github.com/psf/black
target-version = ["py310"]
line-length = 110

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
    | data
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 310
line_length = 110

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true
skip_glob = "data"

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
python_version = "3.10"
pretty = true
show_traceback = true
color_output = true

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true


[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]

[tool.coverage.run]
source = ["tests"]

[coverage.paths]
source = "spire_fyi"

[coverage.run]
branch = true

[coverage.report]
fail_under = 50
show_missing = true